Android: Photo capture implementation with FFmpeg-backend
Image capturing in FFmpeg backend is done by saving frame from
newVideoFrame signal. It should be enough in most cases.

From the other hand, on Android there are features (like always/auto
flash) that can be turn on only when picture is taken. If taking a
picture is done just by saving video frame, we are not able to control
that features. That is why taking picture on Android need to be done
using Camera2 API.

To acheive this assumption, the QAndroidImageCapture was added as
derived class from QFFmpegImageCapture. That allows to call new
takePhoto method on Java side. Taking photo is handled by additional
ImageReader. Additionally we need to control [0]CONTROL_AF_TRIGGER,
to make sure that photo will be taken with correct focus.

[0]https://developer.android.com/reference/android/hardware/camera2/CaptureRequest#CONTROL_AF_TRIGGER

Task-number: QTBUG-113153
Pick-to: 6.6 6.5
Change-Id: I2384b6f5b58d6c9e7e6d76310d5115980679f4ca
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>