Implement initialization of video devices on the first connection
The problem is that VideoDevices are created and managed by
QPlatformMediaIntegration, but signals are connected to
QPlatformMediaDevices.
A possible redesign is here
codereview.qt-project.org/c/qt/qtmultimedia/+/484647,
but the concern is that it's a bit of controversy with the initial idea.

Current patchs solves the poblem directly and covers with UTs.
Idea of the patch: let's handle connection to
QMediaDevices::videoInputsChanged and initialize
QPlatformMediaIntegration and video devices connection.
I propose postpone initialization because users don't expect
us doing any serious magic just on the signal connection.

The UTs were slightly refactored in order to test the case.

Pick-to: 6.6 6.5
Change-Id: I0ca3dfb2126daf69746380fa65096d9d05327f5e
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>