Apply window capture to media capture session
It's a second step of QWindowCapture integration.
The patch propagates window capture to a video sink or
a recorder.

The refactoring has been done with moving some duplicated functionality
to QPlatformVideoSource

The patch is supposed to be integrated partually to 6.5 in order to
have as much code compatibility as possible.

Change-Id: I8c3e6b3a899fb246d48135025ee2213ed9a2f571
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit a7565989a0815e41d4abfde5af5712b2968584b9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>