Refactor internal design for QWindowCapture implementations
The idea is having internal naming ***SurfaceCapture for those
ones that can capture windows and screens, and specific
naming ***ScreenCapture, ***WindoiwCapture.

QPlatformSurfaceCapture is a base class that is able to work with both,
screens and windows. It should be useful since there's a bunch
of similar code for screens a windows copturing.

It's the first step of the refactoring,
other steps are coming soon.

Change-Id: I9b03e4d0b1129f8d39a16195fe3261001eb451f4
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 2bca3087d2b48bed46e6d959b35ebf34d5d256ac)