Doc: Remove/adjust warninglimit .qdocconf variables
The most common limit for the maximum number of allowed documentation
warnings is zero. As all Qt module docs include() a common configuration
from qtbase, it's better to control the base warninglimit from a central
location in qtbase/doc/global.

This allows for a temporary increase of the limit across all modules as
needed - for example, when updating the QDoc binary that the CI
provisions to a version that introduces new types of documentation
warnings.

In qtmultimedia.qdocconf, use the += operator to increase the base
limit to account for known documentation warnings.

Task-number: QTBUG-113326
Change-Id: I2ea30898df41672daf01add4800a401b4294ba77
Reviewed-by: Paul Wicking <paul.wicking@qt.io>