Implement image capturing for ScreenCapture and WindowCapture
The idea is that frames observers, ImageCapture and MediaRecorder,
should communicate with QPlatformVideoSource, not with inheritors.

The approach has been already applied for CaptureSession
and MediaRecorder (in the previous patches).
The current patch applies the approach for ImageCapture with
additional cleanup, e.g. unnecessary stuff, like cameraActiveChange,
has been eliminated.

Pick-to: 6.6 6.5
Change-Id: Ib354b2023cf6716bd1100419cadf888fc5489d97
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>