ffmpeg: Rotate video during playback based on orientation metadata
When playing back video using ffmpeg as multimedia backend, the
video frames are now rotated according to the orientation metadata
stored in the current video stream. Orientation data can e.g. be added
manually to a video file, but is also automatically added by mobile phones
when recording with the phone vertical or upside down.

ffmpeg stores orientation data separate for each stream in a video file
as a "displaymatrix" in a container called "side data". This value
represents a counterclockwise rotation of the video frames, ranging
from -360 to 360 degrees.

With this change, ffmpeg orientation data is converted to a clockwise
rotation angle and stored in QMediaMetaData::Orientation. Values that
doesn't align with 0, 90, 180 or 270 degrees are set to 0.

Each frame gets it rotation angle set in VideoRenderer::renderInternal.

Task-number: QTBUG-113481
Pick-to: 6.5 6.6
Change-Id: I1b83bc30dda03a315844046d12da66075f8cf0b0
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>