Fix x11 surface capture alpha channel
We declare 0xFF alpha channel in formats XRGB, BGRX, etc, but
on some linux systems it's 0xff in xImage.

The patch fixes the problem and suggests SIMD
optimizations for intel platforms that makes a difference
with 4K displays. With the optimizations, the overhead is not so
big comparing with memcpy.

We have a test that covers the case, it failed on some machines.

Pick-to: 6.6 6.5
Change-Id: I78927ea58159ae67bbf993f30b8dec45f617edc1
Reviewed-by: Lars Knoll <lars@knoll.priv.no>