Android: Fix camera image rotation (FFmpeg backend)
The image from the camera was not correctly rotated with ffmpeg backend
(especially for the front camera). To fix the problem, code from
camera-preview documentation[0] was used.

To calculate it correctly, the deviceOrientation value needed to be
changed for Landscape and InvertedLandscape. It needed to be done
because screen orientation is always calculated from
Display.getRotaion(). According documentation[1]:

  "if a device has a naturally tall screen, and the user has turned it
  on its side to go into a landscape orientation, the value returned
  here may be either Surface.ROTATION_90 or Surface.ROTATION_270
  depending on the direction it was turned. The angle is the rotation
  of the drawn graphics on the screen, which is the opposite direction
  of the physical rotation of the device."

In our calculation we need current device orientation - that is why
LandscapeOrientation is treated as 270, while
InvertedLandscapeOrientation is 90.

[0]https://developer.android.com/training/camera2/camera-preview#relative_rotation
[1]https://developer.android.com/reference/android/view/Display#getRotation()

Fixes: QTBUG-114658
Pick-to: 6.6 6.5
Change-Id: I11076f17f89827c982e582e3d8167ac9e4349e17
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>