Fix stuttering video playback and AMD GPU crash on Windows
Video playback on Windows stutters when HW texture conversion is
enabled. This patch works around this issue by disabling HW texture
conversion on Windows by default. By disabling HW texture conversion, we
also avoid a crash situation during video playback with AMD GPUs.

This patch changes the meaning of the
"QT_DISABLE_HW_TEXTURES_CONVERSION" which now has to be non-zero to
disable HW texture conversion on non-Windows platforms. On Windows
platforms, this environment variable must be set, and must have the
value "0" to enable HW texture conversion.

The stuttering playback seems to be related to lack of synchronization
of the frame pool used by the FFMPEG decoder. If the decoder produces
frames faster than they are consumed, frames in this pool may be
overwritten before they are displayed. The mechanism behind the AMD GPU
crash is not yet understood, but could be related to simultaneously
reading from and writing to the same slice in the texture array.

Fixes: QTBUG-113832
Fixes: QTBUG-111543
Fixes: QTBUG-111911
Fixes: QTBUG-111459
Fixes: QTBUG-109213
Pick-to: 6.5 6.6
Change-Id: I12807994ef53644585b132a0da3f38cd3b255e80
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>