Fix screen capture crash if duplication is unavailable
In certain situations, for example when working over remote desktop,
creating the desktop duplication instance may fail. This patch fixes an
error handling issue that led to crash if attempting to start screen
mirroring in this situation.

The fix is to stop relying on asynchronuously calculated frame format,
and instead determine it from the monitor size.

Pick-to: 6.5 6.6
Change-Id: If2359d38f70ee3abc0a2a94b6cf8d4440c6e869a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>