Fix threading problems in QDarwinAudioSink
The most reason of the patch is that renderCallback can be called
by macOS with the same lock as AudioUnitStart/Stop does inside.
That's why we got a deadlock A->B/B->A. The conclusion is that we
shouldn't start/stop the audio unit from our synchronization guard.

Aslo, in some rare cases the renderCallback can be called after stoping,
we should consider it and don't enquire prefetchDevice from it, save
atEnd flag in the main audio sink thread instead.

Aslo, some minor cleanup has been done during fixing things nearby.

Pick-to: 6.6 6.5
Task-number: QTBUG-111567
Change-Id: Ib96d2770f66b83aa1b9a067b51579e3a247de7ff
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>