QtMultimedia: fix duplicated type exports to QML
For some of the QtMultimedia types we need to export both the type as
a value type, and also its enums into a namespace.
With the pre-existing approach this results into duplicated entries
in the qmltypes file, which confuses Qt Design Studio.

Fix it by introducing a dummy derived class for each of such types, and
using this derived type to export the enums.

This patch fixes exports for the following classes:
* QAudioDevice
* QCameraDevice
* QMediaFormat
* QMediaMetaData

Fixes: QTBUG-115360
Fixes: QTBUG-115563
Fixes: QTBUG-115564
Fixes: QTBUG-115566
Change-Id: I9073ec8317e1ef4f5f2a613406c750f1d1cd6d4e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 16ab1dd1d13b6644476e2bc33ac9cb2310903c32)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>