Fix mediaplayer scroll issue if stream duration is missing
If a media file is missing per stream duration data, the mediaplayer
slider operations to scroll in video would restart the video from start.
The cause of the issue was that the mediaplayer duration was calculated
from the stream durations and became 0.

The ffmpeg demuxer already has more advanced heuristics to calculate the
media file duration, and this patch fixes the issue by also taking the
media file duration into account.

Fixes: QTBUG-115747
Pick-to: 6.5 6.6
Change-Id: I906f62c15ffdccadf1da07a75a109961d897b1b0
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>