Fix application hang when camera is unplugged
If a camera device is unplugged from the system and then plugged back
in application will hang during the attempt to free resources acquired
for the previous device. Destruction of the camera object hangs on a
semaphore that is never released if camera communication is lost. The
fix is to check for error conditions before attempting to acquire the
semaphore.

Task-number: QTBUG-115463
Task-number: QTBUG-116467
Change-Id: I181ae9339ce3c903a126d8fbb21ebe384fb124a5
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 084ea0f0705c6d49b4da4eb10de980a7cbf2b797)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>