Android-ffmpeg: handle maxImages acquired problem
It was possible that camera was crashed with IllegalStateException:

  "maxImages (10) has already been acquired, call #close before
  acquiring more."

The problem is described in documentation[0]. To avoid this, it is
recommended to call Image#close in the meantime, which will exhaust
the underlying queue. But in Java code, we don't know if the frame
is processed by ffmpeg or not. That is why, images are closed in C++
code (in QAndroidCameraFrame destructor).

There may still be some delay in the thread that will cause the
mentioned exception. Especially when launching or bringing an app to the
foreground. That is why IllegalStateException need to be handled.

[0]https://developer.android.com/reference/android/media/ImageReader#acquireLatestImage()

Pick-to: 6.5
Fixes: QTBUG-115460
Fixes: QTBUG-116526
Fixes: QTBUG-114932
Change-Id: I2d2252859903e64111d681e1eca403516374b0a2
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit c5e5d619107568050d9857d5149bd1b7558b904b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>