Set AV_HWDEVICE_TYPE_CUDA first priority on Linux
The reason is that on nvidia graphic cards AV_HWDEVICE_TYPE_VAAPI
gets failed on the vaapi hw format selection, and it falls back
to sw decoding.

Let's go with AV_HWDEVICE_TYPE_CUDA first on nvidia cards;
on other cards, nvidia drivers cannot be detected and cuda is not used.

Pick-to: 6.6 6.5
Change-Id: I605f5525334ecb06198a8e1e410e637e32a8ee32
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>