MediaRecorder: Fix crash if output path contains whitespace
If a URL containing whitespace characters was specified as
outputLocation in MediaRecorder, the application crashed when
recording was started.

This is because we converted the path using  QUrl::toEncoded(),
which converts whitespaces to %20, which in turn leads to a
segmentation fault in the ffmpeg backend. The soulution is to use
QUrl::toString(QUrl::PreferLocalFile) instead, which returns a
QString formatted as a local file path, preserving whitespaces.

To make things more readable, the resolving of outputLocation to
a valid file path is also moved out of Encoder::Encoder. This
constructor now receives QString &filePath instead of a QUrl.
Resolving the outputLocation to a valid file path and converting
it to QString is now done inside QFFmpegMediaRecorder::record,
with help from  QMediaStorageLocation::generateFileName. Some
other small simplifications and name tweaks was also made.

Fixes: QTBUG-116801
Pick-to: 6.5
Change-Id: Ia4b3befa99c639185bec29382cb18c0cd7e3c23e
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit b3a1c56ab98c6eedd68288e043fb55e8bfd9fe6c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>