JNI: replace TYPE declarations with CLASS declarations
That we have two macros to declare a C++ type to represent a Java class
is confusing. The TYPE macro as of now allows us to declare array types,
but with QJniArray we won't need that anymore, and can just use Class[]
as the type instead. Changing that will be a follow-up commit; for now,
get rid of TYPE-usages to declare regular classes.

Change-Id: Id75f6d653a75fc8f6d69e3825302a525928df792
Reviewed-by: Petri Virkkunen <petri.virkkunen@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>