Test QVideoFrame::map also with the YUV422P format
The QVideoFrame::map function was not tested with the YUV422P format.
This reveals incorrect calculation of color plane strides and offsets.

The YUV422P format is a planar format, where the Y, U, and V components
are in separate color planes. The U and V planes are sub-sampled
horizontally to half the width of the Y plane, but have the same height
as the Y plane.

                +----------- Y_stride ------------+
                |                                 |
                |            Y (8 bit)            | height
                |                                 |
 U_offset  -->  +--- U_stride ---+----------------+
                |                |
                |   U (8 bit)    | height
                |                |
 V_offset  -->  +--- V_stride ---+
                |                |
                |   V (8 bit)    | height
                |                |
                +----------------+

Here, we calculate the U and V stride incorrectly as

    U_stride = Y_stride / 4 (Wrong)

It should have been

    U_stride = Y_stride / 2 (Correct)

Also, the V offset is incorrectly calculated as if the U plane was sub-
sampled both in width in height:

    V_offset = Y_stride * height + Y_stride / 2 * height / 2 (Wrong)

It shouldhave been

    V_offset = Y_stride * height + Y_stride / 2 * height (Correct)

Task-number: QTBUG-116991
Pick-to: 6.5 6.6
Change-Id: I97e36bf1b7148aa4aed69060d0d0146713edd370
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>