Fix broken QVideoFrame::map with YUV422P pixel format
A regression in change-344586 caused the QVideoFrame::map function
to set the incorrect strides for the YUV422P format. This caused
QVideoFrame::bytesPerLine to return incorrect stride value, and thereby
giving wrong sizes for frame allocations that disagree with FFmpeg.

The YUV422P format is a planar format, where the Y, U, and V components
are in separate color planes. The U and V planes are sub-sampled
horizontally to half the width of the Y plane, but have the same height
as the Y plane.

                +----------- Y_stride ------------+
                |                                 |
                |            Y (8 bit)            | height
                |                                 |
 U_offset  -->  +--- U_stride ---+----------------+
                |                |
                |   U (8 bit)    | height
                |                |
 V_offset  -->  +--- V_stride ---+
                |                |
                |   V (8 bit)    | height
                |                |
                +----------------+

After this fix, the U and V strides are correctly calculated as

    U_stride = Y_stride / 2

and

    U_stride = Y_stride / 2

Thus we get the correct plane size as

    V_offset = Y_stride * height + Y_stride / 2 * height

Fixes: QTBUG-116991
Pick-to: 6.5 6.6
Change-Id: Ifd9f83ee75e0d5f869675232d13d7e2b84a04c7b
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>