Add audio warm-up opt out option
Current workaround to fix audio cutting out in the beginning of the
audio on Windows introduced some weird side effects which are hard to
reproduce, such as: blocking sleep mode, stucking CPU core at the
minimum frequency. So far, there's no any other solution to the cutting
sound issue, so an environment variable that provides the ability to
opt out is introduced.

Task-number: QTBUG-117528
Task-number: QTBUG-117612
Pick-to: 6.6 6.5
Change-Id: I8b6ecab5cb3ae9136fd29adf5186d15c9128707f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>