Use aggregation instead of inheritance with FFmpeg MediaDataHolder
Playback engine was inheriting MediaDataHolder in a 'uses' relationship, not in
an 'is a' relationship. In this case, aggregation is preferable to adhere to the
single responsibility principle and to make dependencies more explicit in the
code.

Change-Id: I4342aa6e0c2c47e074eb88d20c73d630315f573c
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 7a30b8651537b21bd5eadda7509b8e8029928a28)
(cherry picked from commit 563eab133079f5f73f0d2072cc748496c7593442)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>