Move initialization code to media data holder constructor
The media data holder was partially constructed using a member function.
We can use the constructor instead to make it clear that this is indeed
construction.

Now, the media data holder is created by a 'create' factory function
that first prepares the AVFormatContext object, and if successful
creates a media data holder using its constructor. By doing this, it is
a bit clearer where the AVFormatContext object is being accessed.

Pick-to: 6.5 6.6
Change-Id: I0e0f396c7f7bdec936dcb4411091912bad90bbff
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>