Keep UI responsive while loading media
The mediaplayer caused the UI thread to become unresponsive while
loading media. This was particularly noticeable when loading network
streams, especially when connection could not be established.

The cause of the issue was that loading the file or connecting to
network was done on the UI thread. If the network resource was
unavailable, the UI thread would hang for the duration of the socket
connect timeout (5 seconds).

This patch fixes the issue by offloading media loading to a worker
thread. This is thread safe because MediaDataHolder::create is a static
function without any shared data. Only MediaDataHolder::create is
required to run on a separate thread.

Once the media is loaded, the media player is notified, and playback can
start. An extra state variable is introduced to be able to keep track of
play/pause/stop events that arrive while media is loaded in the
background.

Cancellation is not yet implemented. The result is that app shutdown and
loading another URL while the first is being processed will still cause
UI thread hang for a duration determined by the network timeout (5
seconds).

Task-number: QTBUG-116779
Pick-to: 6.5
Change-Id: I76c8d08c58d2f7795b3e3828a070b7695336978a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 4b51aa786a0a75dc2f193a2d2a16080fc12ec24f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>