Add network connect cancellation support with FFmpeg media player
If attempting to connect to an unresponsive network resource, the media
player will block the UI thread for the duration of the configured
network timeout (5s) if the media player is destroyed (application
closed) or if a second attempt is made to access another network
resource within the timeout.

This patch fixes this issue by leveraging existing FFmpeg mechanism for
interrupting ongoing network connect attempts. Cancellation is supported
in two situations:
1. The media player is destroyed while attempting to connect to an
   unresponsive network resource.
2. setMedia is called while a previous call to setMedia is still in
   progress towards an unresponsive network resource.

Fixes: QTBUG-116779
Pick-to: 6.5 6.6
Change-Id: I84c151a7990adbe03623957ed9d5e195d2432388
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>