Use xioctl function wrapper for v4l2 camera
ioctl documentation says:
"On error, -1 is returned, and errno is set to indicate the error."
So we have an undefined state when result is -1 but errno is not set.

xioctl tries ioctl again in the undefined state,
and it is used in v4l2 camera example, see
https://www.kernel.org/doc/html/v4.9/media/uapi/v4l/capture.c.html

This suspects that ioctl commands, that have been sent to v4l2,
may occasionally return this state and should be retried again.

Pick-to: 6.5
Change-Id: I738990855886d5628834f0e5934a13359650261b
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 2011f29aabe75c2e79029b85acdcaeba3a0284c1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>