Refactor QML Camera example orientation and control layout state logic
In QML Camera example, the state of the photo/video controls that
determines the layout those controls was updated by the setState
function. This function was called on camera UI width change, and
initially only because of stillControls.state being bound to the
function. The videoControls.state was then bound to stillControls.state.
This caused qmllint warnings (unqualified access to setState() and info
message because setState() doesn't return a value) and log spamming
whenever the window width changed.

The orientation and control layout state logic is now encapsulated in a
separate Item "controlLayout", and the stillControls and videoControls
objects both bind to the state and buttonsWidth properties of this new
Item. Additionally the message about state change is logged only when
the state actually changes.

Documentation is also updated to display the correct part of the code
in the section that mentions the control layout.

Task-number: QTBUG-113682
Pick-to: 6.5 6.6
Change-Id: I99e5c0c3d3bb369dd985cd06fdb51d687ef24f1f
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>