Refactor QML Camera example orientation and control layout state logic
Manually cherry-picked from commit
2785589db4447d3ea421ba59130d6be4e6574b6b

The documentation modifications in the original commit was not
included, because the affected paragraphs wasn't part of the 6.2 docs.

Description below is copied from original commit:
-----

In QML Camera example, the state of the photo/video controls that
determines the layout those controls was updated by the setState
function. This function was called on camera UI width change, and
initially only because of stillControls.state being bound to the
function. The videoControls.state was then bound to stillControls.state.
This caused qmllint warnings (unqualified access to setState() and info
message because setState() doesn't return a value) and log spamming
whenever the window width changed.

The orientation and control layout state logic is now encapsulated in a
separate Item "controlLayout", and the stillControls and videoControls
objects both bind to the state and buttonsWidth properties of this new
Item. Additionally the message about state change is logged only when
the state actually changes.

Task-number: QTBUG-116315
Change-Id: I99e5c0c3d3bb369dd985cd06fdb51d687ef24f1f
(cherry picked from commit 2785589db4447d3ea421ba59130d6be4e6574b6b)
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>