Clarify section on video compression patents
Use of patented video compression standards such as H.264 may require
payment of royalty fees to the patent holders, and it is up to the
service provider to understand if such fees may apply to their product.

This patch clarifies that such royalties may apply independently on the
implementation being used. Notably, it is not tied to FFmpeg or
any of the backends provided by Qt, and may apply even if the service is
provided through an operating system service or system call.

Task-number: QTBUG-118195
Pick-to: 6.5
Change-Id: If6988c9985de9f74e3eee6d1067ccdd7fef50531
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 62f0b382ddd5c04674ea2e2252b68a13c9e4ded8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>