Improve mediaplayer test naming and split some big tests
This patch attempts to improve naming of some of the mediaplayer tests
to make it easier to see what the test is covering. In addition, doc
strings are added to better describe some of the tests, and some long
scenario tests are split into smaller ones to ease maintainability.
Some tests are also extended to cover more of the media player's state,
to better specify which function should change which properties.

The goal of this patch is to make gaps in test coverage easier to spot,
without reducing test coverage.

Task-number: QTBUG-116779
Pick-to: 6.5
Change-Id: I1d6e1693e4b3d909b3ecef608cea22134f16f526
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 40356de19a07af4b88cfd698850d482fd0929060)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>