Simplify lifetime handling of IMFSample instances even more
In this change we replace a queue of IMFSample interface pointers with
a queue of smart pointers in the SamplePool class. This way, we turn the
queue into an owning queue, and reduce the need for manually calling
AddRef and Release on the contained COM interfaces.

Pick-to: 6.5
Change-Id: I10f00855f94005e985f9a83326890c78c8634590
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit b295b244bd4ff4c3bbea298664830c1c54c068f1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>