Simplify lifetime handling of IMFSample instances for frame stepping
In this change we replace a queue of IMFSample interface pointers with
a queue of smart pointers in the FrameStep struct. This way, we turn the
queue into an owning queue, and reduce the need for manually calling
AddRef and Release on the contained COM interfaces.

Pick-to: 6.5
Change-Id: I3f284b9fb636b6284342f9ccce6d7927bab94ba8
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
(cherry picked from commit fde9561d9a73306fe70e1679e3ceee847ba452da)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>