Simplify lifetime handling when creating EVR video samples
In this change, we reduce the need for manual handling of COM interface
reference counting when creating video samples in EVR MediaPlayer.

This is done by maintaining the interface pointers in a list of smart-
pointers, which takes care of the reference counting for us.

Pick-to: 6.5
Change-Id: I119203ba1aae657b5ddd389f306762454933ff46
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 8cd1fbf6a790d1c8ad986920949405e449d16098)