Add Qt Multimedia attributions to "Licenses Used in Qt" page
This patch adds an attribution to FFmpeg which is used by QtMultimedia.
Note that Qt Multimedia's landing page already had an attribution
section. This content is now moved into the qt_attribution.json file for
FFmpeg. The QtMultimedia landing page is simplified and refers to the
attributions instead of going into the licensing details.

This patch has one weakness. The automatic attribution generation adds a
path to where the sources can be found. On the ffmpeg attribution page,
it will state

    "The sources can be found in qtmultimedia/src/3rdparty/ffmpeg."

This is not true, because the ffmpeg sources are not part of the Qt
source tree, and are not located in the 3rdparty/ffmpeg directory.

It is hard to find an appropriate copyright string for FFmpeg from their
documentation. However, when running ffmpeg.exe, the text

    "Copyright (c) 2000-2023 the FFmpeg developers"

shows up. This is used as a copyright string in this patch.

The licensing terms for FFmpeg refers to use of libjpeg in FFmpeg. It is
not possible to find which version of libjpeg is being used by FFmpeg,
or what modifications have been made. We therefore just add the general
Independent JPEG Group License.

Change-Id: Ie8eba1fef0ac3b2e4283f7d88c20163eee49763e
Task-number: QTBUG-118734
Fixes: QTBUG-118204
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit 335ee775db270158455d9fbf89db5c7671003d8a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit adfd9d721bba126eb2d27fa5d5a39d5b97d7c515)