Harden video processing pipeline for out-of-range signals
This patch makes the video processing pipeline well-behaved when facing
out-of-range video signals. It replaces gamma 2.2 approximation of sRGB
with a proper sRGB transfer function as described in
https://en.wikipedia.org/wiki/SRGB, and saturates output RGB signal to
be within the range of the underlying texture type. Together, these two
changes prevents Infinity or NaN values in video output, and prevents
potential wrap-around with out-of-range values.

Conversion from linear space to sRGB using an inverse gamma 2.2 transfer
function is problematic because x^(1/2.2) is imaginary for negative x.
In particular, the openGL pow(x, y) is undefined for negative x. The new
equations are linear for low input values, and therefore well behaved
with negative input. Now that we have well-defined RGB output also with
out-of-range values, we can safely clamp any out-of-range value to be
within the expected [0.0, 1.0] range.

Out-of-range inputs can be caused by numerical inaccuracies when
converting from YUV to RGB, and from out-of-range values in the input
YUV video signal. In particular with 'video' range media, values outside
the 16-235 range is expected and must be handled gracefully.

Fixes: QTBUG-113317
Pick-to: 6.5 6.6 6.7
Change-Id: Ia62777650efc19abf7e6f2a2a65f20deb0ec2a1c
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>