QSGVideoMaterial: get rid useless mutex
We can remove the useless m_frameMutex from QSGVideoMaterial,
since both functions it guards are always called from the same
thread (QSGRenderThread). It's also non-sense to guard only these
functions, since m_currentFrame can be accessed from other functions
such as QSGVideoMaterialRhiShader::updateUniformData(),
QSGVideoNode::setTexturedRectGeometry(), etc. which are not guarder by
the mutex. But this is not a problem, because they are also called from
QSGRenderThread.

Pick-to: 6.6 6.5
Change-Id: Iae2d8c4f8b16f10faed51b205513350c793dd1b1
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>