Delay emitting playingChanged event until after state has changed
The QMediaPlayer was emitting playingChanged event before changing its
state. Clients would therefore read the old value of isPlaying when
responding to the playingChanged signal.

This patch updates the internal media player state before emitting the
playingChanged signal, to ensure that clients get the current value when
responding to the signal.

Fixes: QTBUG-117006
Pick-to: 6.6 6.5
Change-Id: I9d45d6e0613ff901f4a02759798f2e25386b6c64
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>