Doc update on negative playback rates, and hardened FFmpeg handling
The FFmpeg media player does not support negative playback rates, even
if the documentation stated that negative rates can be used.

This is fixed by updating the documentation to state that negative rates
are not supported (independently of backend). Note: Negative playback
rates may still work with some backends, but it is not considered
supported (or maintained).

In addition, the FFmpeg backend has been changed to clamp negative
playback rates to 0, instead of asserting.

This patch also fixes an issue where playbackRateChanged signals were
not emitted by the FFmpeg backend. The mediaplayer example is updated
to make this issue easier to spot in the future.

Task-number: QTBUG-118654
Pick-to: 6.6 6.5
Change-Id: Iaaffa9fa39bdd462ad5b5a838390325d78d64e85
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>