Doc: Fix template information for QAudioBuffer members
When QDoc parses a project, it parses the source code to extract the
user-provided documentation and perform sanity checkings based on the
code itself on it.

When QDoc parses an "\fn" command as part of this process, it tries to
understand, based on its intermediate representation built on the
information extracted from the code-base, which "documentable element"
the "\fn" refers to.

When QDoc performs this "matching" process, it takes into consideration
only a certain amount of information.
For example, no checking is performed over the template declaration of a
callable.

Due to some upcoming documentation, where two callables are
indistinguishable to the current process, as they differ only in their
template declaration, QDoc will start to take into consideration the
template declaration of a callable when matching.

This implies that an "\fn" command should now provide information
parity, with regards to template declaration for callables, with the
code-base so that QDoc can perform the match correctly.

The documentation for some of the members of `QAudioBuffer` is not in sync
with the intended target template declaration.

In particular, `QAudioBuffer` provides a public templated version and a
private non-templated version of `data`/`constData`.

The documentation for those member was provided as an auto-tying block
that was intended to tie to the private non-templated members, while
documenting both the private and non-private API.

As the private and public version only differ in their template
declaration, which was not previously considered, and in their return
type, which is correctly ignored when matching, QDoc would match the
documentation to the public templated version, even if in the code it
was positioned to match the private non-templated version.

Now that QDoc will consider the template information, the matching will
correctly pick the private non-templated version of the methods, thus
producing no documentation in the final output as private methods are
considered internal.

To fix the issue, the auto-tying documentation blocks for
`data`/`constData` were modified to use an `\fn` topic pointing to the
public templated version.

The body of the documentation was modified to avoid implying the
presence of two overloads for the methods, instead referring directly to
the public templated version, as the private version should not be
generally referred to in the final output.

Some spacing between the documentation blocks for `data`/`constData` and
the out-of-line definition of the private non-templated version of the
methods was added in the implementation file, so that it would be easier
to see that the documentation blocks are not intended to auto-tie to the
following definition.

A commment referring to QDoc inability to recognize both the templated
and non-templated version of the API was removed as it is soon going to
be out-of-date and is not relevant to the presented version of the
documentation anymore.

Task-number: QTBUG-118080
Change-Id: I2280098f76fd34567db8ec588615621643329fd0
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>