Add tests for multimedia rendering pipeline
Video playback with QtMultimedia depends on several parts to get the
colors right, including color space transformation matrices and
transformations in shaders. It is easy to inadvertedly introduce
regressions in this part of our codebase.

This patch introduces regression testing of the color transformations
applied when displaying video content in QtMultimedia.

Note that this patch reveals that videos encoded as ICM2 have their U
and V values swapped, as evident from the test images.

Change-Id: Ibd156142fcc267e845236ae6562a7cb8ebd72baa
Pick-to: 6.5 6.6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>